fix: removing warnings
Mainly warnings of precision loss.
In Text, correcting a miss ")".

+++ src/Sprite.cpp
@@ -32,7 +32,7 @@ void Sprite::SetFrame(int frame) {
 void Sprite::SetFrameCount(int frame) {
     frameCount = frame;
     currentFrame = 0;
-    associated.box.w = width/frameCount; 
+    associated.box.w = ((double)width)/frameCount; 
 }
 
 void Sprite::SetFrameTime(float frameTime) {
@@ -52,7 +52,7 @@ void Sprite::Open(std::string file) {
         exit(-1);
     }
     SetClip(0,0,width/frameCount,height);
-    associated.box.w = width/frameCount;
+    associated.box.w = ((double)width)/frameCount;
     associated.box.h = height;
 }
